+++ netcomms/pages/boards_utils/board_joining.go
@@ -18,7 +18,7 @@ func JoinBoardPage(w http.ResponseWriter, r *http.Request) {
 		vars := mux.Vars(r)
 		boardID, _ := strconv.Atoi(vars["id"])
 		tmpl, _ := template.ParseFiles("./static/boards_utils/board_joining.html")
-		b, _ := all_boards.BoardByID(boardID)
+		b, _ := all_boards.BoardByID(uint64(boardID))
 		tmpl.Execute(w, b)
 	}
 }
@@ -31,7 +31,7 @@ func ProcBoardJoining(w http.ResponseWriter, r *http.Request) {
 		inpPwd := r.PostForm.Get("pwd")
 		vars := mux.Vars(r)
 		boardID, _ := strconv.Atoi(vars["id"])
-		if all_boards.AddObserver(boardID, account_logic.GetUserID(r), inpPwd) {
+		if all_boards.AddObserver(uint64(boardID), account_logic.GetUserID(r), inpPwd) {
 			http.Redirect(w, r, fmt.Sprintf("/board%d", boardID), http.StatusSeeOther)
 		}
 	}
